Fix triggering vote toggle when pasting or typing search terms.
When pasting using ctrl+v, it is common for ctrl to be released prior to releasing v. This resulted in false triggers due to uncancelled keyup events in the contenteditable rich textbox controls. This is fixed by using the keydown event instead of the keyup event. This way, the absence of modifier keys can be correctly detected.

When typing into the search textbox, the letter v was triggering a vote incorrectly. This is due to the keyup event for simpler inputs (non-contenteditable ones) not being marked cancelled by Jira. This is fixed by allowing the default actions of the keydown event to run and detecting the lack of a beforeinput event prior to proceeding as this event will fire in contexts where the input is treated as an input character both for normal inputs and contenteditable. Another facet of this fix is to specifically test for the absence of a select ancestor on the event target as select elements don’t receive beforeinput events.

Fixes #1.